modularize collection view for both my cards and community pages
