+++ components/PlayerCard/PlayerCard.styles.ts
@@ -1,10 +1,9 @@
 import { makeStyles } from '@material-ui/core'
 
 const useStyles = makeStyles((theme) => ({
-  container: {
-    display: 'flex',
-    flex: '1 1 auto',
-    flexDirection: 'column',
+  cardContainer: {
+    maxWidth: '100%',
+    maxHeight: '100%',
   },
 }))
