+++ pages/Community/[user].tsx
@@ -1,11 +1,14 @@
 import React from 'react';
 import { useRouter } from 'next/router';
+import CollectionView from '../../components/CollectionView/CollectionView';
 
 const CommunityMemberPage = () => {
   const router = useRouter()
   const { user } = router.query;
 
-  return <h1>{user}'s Cards</h1>;
+  return (
+    <CollectionView />
+  );
 }
 
 export default CommunityMemberPage;
